requirements.txt
pyaml==21.10.1
PyYAML==6.0
requests==2.28.2